Reintroduce invoice contract, useful for gateways
